date picker under one component
