+++ libs/expo/shared/ui-components/src/lib/index.ts
@@ -26,7 +26,6 @@ export {
   parseToDate,
 } from './Date';
 export { default as DatePicker } from './DatePicker';
-export { default as DatePickerV2 } from './DatePicker_V2';
 export { default as DeleteModal } from './DeleteModal';
 export { default as DirectionsPopup } from './DirectionsPopup';
 export { default as Divider } from './Divider';
